+++ settings.gradle.kts
@@ -2,7 +2,7 @@ rootProject.name = "registration-service"
 
 include(":launcher")
 include(":extensions:registration-service")
-include(":extensions:registration-service-mocks")
+include(":extensions:registration-policy-region-any")
 include(":extensions:participant-store-spi")
 include(":extensions:dataspace-authority-spi")
 include(":system-tests")
